+++ src/services/@l13/fse.ts
@@ -59,19 +59,31 @@ export function walktree (cwd:string, options:Callback|Options, callback?:Callba
 
 export function unlinkSync (pathname:string) :number {
 	
-	const stat = fs.lstatSync(pathname);
+	const stat = lstatSync(pathname);
 	let total = 0;
 	
-	if (stat.isDirectory()) {
-		fs.readdirSync(pathname).forEach((name) => total += unlinkSync(path.join(pathname, name)));
-		fs.rmdirSync(pathname);
-		total++;
-	} else if (stat.isFile() || stat.isSymbolicLink()) {
-		fs.unlinkSync(pathname);
-		total++;
+	if (stat) {
+		if (stat.isDirectory()) {
+			fs.readdirSync(pathname).forEach((name) => total += unlinkSync(path.join(pathname, name)));
+			fs.rmdirSync(pathname);
+			total++;
+		} else if (stat.isFile() || stat.isSymbolicLink()) {
+			fs.unlinkSync(pathname);
+			total++;
+		}
 	}
 	
 	return total;
+	
+}
+
+export function lstatSync (pathname:string) {
+	
+	try {
+		return fs.lstatSync(pathname);
+	} catch (error) {
+		return null;
+	}
 	
 }
